refactor(Components): extracted
