+++ android/app/src/dev/res/xml/network_security_config.xml
@@ -6,6 +6,7 @@
         <domain includeSubdomains="false">10.0.3.2</domain>
         <domain includeSubdomains="false">192.168.86.21</domain>
         <domain includeSubdomains="false">192.168.2.21</domain>
+        <domain includeSubdomains="false">192.168.2.25</domain>
         <domain includeSubdomains="false">192.168.0.101</domain>
         <domain includeSubdomains="false">192.168.1.101</domain>
     </domain-config>
